* added overflow style to final results panel * moved button "next sportsman" to the middle panel * add new confirm dialog of type message * when competition ends show message "thank you" * support that user can have multiple user types

+++ controllers/login.js
@@ -1,28 +1,58 @@
-app.controller("loginController", function ($scope, $http, $window, $location, $rootScope, constants,SocketService) {
+app.controller("loginController", function ($scope, $http, $window, $location, $rootScope, constants, SocketService, confirmDialogService) {
     $scope.regex = constants.regex;
 
-    $scope.submit = function(isValid){
-        if(isValid){
-            var user ={
-                userID : $scope.username,
-                password : $scope.password};
-            $http.post(constants.serverUrl+ '/login', user)
-                .then(function (response) {
-                        $window.sessionStorage.setItem('name', response.data.firstname);
-                        $window.sessionStorage.setItem('id', response.data.id);
-                        $rootScope.name = $window.sessionStorage.getItem('name');
+    $scope.submit = function (isValid) {
+        if (isValid) {
+            var user = {
+                userID: $scope.username,
+                password: $scope.password
+            };
+            $http.post(constants.serverUrl + '/loginFirstStep', user)
+                .then(function (firstResponse) {
+                    confirmDialogService.chooseUserType(firstResponse.data.dbResults.usertype, (userType) =>{
+                        firstResponse.data.dbResults.usertype = userType;
+                        $http.post(constants.serverUrl + '/loginSecondStep', firstResponse.data)
+                            .then(function (response) {
+                                    $window.sessionStorage.setItem('name', response.data.firstname);
+                                    $window.sessionStorage.setItem('id', response.data.id);
+                                    $rootScope.name = $window.sessionStorage.getItem('name');
 
-                        $window.sessionStorage.setItem('access', response.data.access);
-                        $rootScope.access = $window.sessionStorage.getItem('access');
-                        $window.sessionStorage.setItem('isFirstLogin', response.data.isFirstTime);
-                        $window.sessionStorage.setItem('token', response.data.token);
-                        SocketService.emit('login',{ userId: user.userID });
-                        $location.path('/home');
-                    },
-                    function (error) {
-                        $scope.isError = true;
-                        console.log(error.data);
-                    });
+                                    $window.sessionStorage.setItem('access', userType);
+                                    $rootScope.access = $window.sessionStorage.getItem('access');
+                                    $window.sessionStorage.setItem('isFirstLogin', response.data.isFirstTime);
+                                    $window.sessionStorage.setItem('token', response.data.token);
+                                    SocketService.emit('login', {userId: user.userID});
+                                    $location.path('/home');
+                            }, function (err) {
+                                console.log(err)
+                            });
+                    })
+                }, function(err){
+                    console.log(err)
+                });
+
+
+            // $http.post(constants.serverUrl + '/login', user)
+            //     .then(function (response) {
+            //             confirmDialogService.chooseUserType(response.data.access, (userType) => {
+            //                 $window.sessionStorage.setItem('name', response.data.firstname);
+            //                 $window.sessionStorage.setItem('id', response.data.id);
+            //                 $rootScope.name = $window.sessionStorage.getItem('name');
+            //
+            //                 $window.sessionStorage.setItem('access', userType);
+            //                 $rootScope.access = $window.sessionStorage.getItem('access');
+            //                 $window.sessionStorage.setItem('isFirstLogin', response.data.isFirstTime);
+            //                 $window.sessionStorage.setItem('token', response.data.token);
+            //                 SocketService.emit('login', {userId: user.userID});
+            //                 $location.path('/home');
+            //                 $scope.$apply()
+            //             });
+            //
+            //         },
+            //         function (error) {
+            //             $scope.isError = true;
+            //             console.log(error.data);
+            //         });
         }
     };
 });
